refactor: [FE] 로고 이미지 성능 개선
- Intent
캐싱되어 재사용될 수 있도록 기존 두가지 크기로 사용되는 로고 이미지를 더 작은 용량에 큰 사이즈를 가진 한가지 이미지로 변경했어요. 라이트하우스 안내를 따라 레이아웃 계산을 줄이기 위해 가로세로 크기를 지정했어요.
